requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pythainlp>=4.0.0
tqdm
accelerate
bitsandbytes
peft
